rename to range and set views as protected
